Moves the checkbox styles to input-checkbox-styles.js
