+++ components/inputs/input-checkbox.js
@@ -1,5 +1,6 @@
 import '../expand-collapse/expand-collapse-content.js';
 import '../tooltip/tooltip.js';
+import { _generateInputCheckboxStyles, cssSizes } from './input-checkbox-styles.js';
 import { css, html, LitElement, nothing } from 'lit';
 import { classMap } from 'lit/directives/class-map.js';
 import { FocusMixin } from '../../mixins/focus/focus-mixin.js';
@@ -8,116 +9,9 @@ import { getUniqueId } from '../../helpers/uniqueId.js';
 import { ifDefined } from 'lit/directives/if-defined.js';
 import { InputInlineHelpMixin } from './input-inline-help.js';
 import { offscreenStyles } from '../offscreen/offscreen.js';
-import { registerSemanticVariableForSvgImageUrl } from '../colors/colors.js';
 import { SkeletonMixin } from '../skeleton/skeleton-mixin.js';
 
-registerSemanticVariableForSvgImageUrl(
-	'--d2l-input-checkbox-check-image',
-	`<svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24">
-		<path fill="var(--d2l-theme-icon-color-standard)" d="M8.4 16.6c.6.6 1.5.6 2.1 0l8-8c.6-.6.6-1.5 0-2.1-.6-.6-1.5-.6-2.1 0l-6.9 7-1.9-1.9c-.6-.6-1.5-.6-2.1 0-.6.6-.6 1.5 0 2.1l2.9 2.9z"/>\
-	</svg>`
-);
-
-registerSemanticVariableForSvgImageUrl(
-	'--d2l-input-checkbox-indeterminate-image',
-	`<svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24">
-		<path fill="var(--d2l-theme-icon-color-standard)" d="M7.5,11h9c0.8,0,1.5,0.7,1.5,1.5l0,0c0,0.8-0.7,1.5-1.5,1.5h-9C6.7,14,6,13.3,6,12.5l0,0C6,11.7,6.7,11,7.5,11z"/>
-	</svg>`
-);
-
-export const cssSizes = {
-	inputBoxSize: 1.2,
-	checkboxMargin: 0.5,
-};
-
-export const checkboxStyles = css`
-	input[type="checkbox"].d2l-input-checkbox {
-		--d2l-input-checkbox-background-image: none;
-		--d2l-input-checkbox-background-color: var(--d2l-theme-background-color-interactive-faint-default);
-		--d2l-input-checkbox-background-image-disabled:
-			linear-gradient(
-				var(--d2l-theme-background-color-interactive-faint-disabled),
-				var(--d2l-theme-background-color-interactive-faint-disabled)
-			),
-			var(--d2l-input-checkbox-background-image);
-		-webkit-appearance: none;
-		-moz-appearance: none;
-		appearance: none;
-		background-image: var(--d2l-input-checkbox-background-image);
-		background-position: center center;
-		background-repeat: no-repeat;
-		background-size: ${cssSizes.inputBoxSize}rem ${cssSizes.inputBoxSize}rem;
-		border-radius: 0.3rem;
-		border-style: solid;
-		box-sizing: border-box;
-		display: inline-block;
-		height: ${cssSizes.inputBoxSize}rem;
-		margin: 0;
-		padding: 0;
-		vertical-align: middle;
-		width: ${cssSizes.inputBoxSize}rem;
-	}
-	input[type="checkbox"].d2l-input-checkbox:checked {
-		--d2l-input-checkbox-background-image: var(--d2l-input-checkbox-check-image);
-	}
-	input[type="checkbox"].d2l-input-checkbox:indeterminate {
-		--d2l-input-checkbox-background-image: var(--d2l-input-checkbox-indeterminate-image);
-	}
-	input[type="checkbox"].d2l-input-checkbox,
-	input[type="checkbox"].d2l-input-checkbox:hover:disabled {
-		background-color: var(--d2l-input-checkbox-background-color);
-		border-color: var(--d2l-theme-border-color-emphasized);
-		border-width: 1px;
-	}
-	input[type="checkbox"].d2l-input-checkbox:hover:disabled {
-		border-color: var(--d2l-theme-border-color-disabled);
-	}
-	input[type="checkbox"].d2l-input-checkbox:hover,
-	input[type="checkbox"].d2l-input-checkbox:focus,
-	input[type="checkbox"].d2l-input-checkbox.d2l-input-checkbox-focus,
-	:host(.d2l-hovering) input[type="checkbox"]:not(:disabled).d2l-input-checkbox {
-		border-color: var(--d2l-input-checkbox-border-color-hover-focus, var(--d2l-theme-border-color-focus));
-		border-width: 2px;
-		outline: none;
-	}
-	input[type="checkbox"].d2l-input-checkbox:disabled,
-	input[type="checkbox"].d2l-input-checkbox:where([aria-disabled="true"]) {
-		background-image: var(--d2l-input-checkbox-background-image-disabled);
-		border-color: var(--d2l-theme-border-color-disabled);
-	}
-	@media (forced-colors: active) {
-		input[type="checkbox"].d2l-input-checkbox:checked,
-		input[type="checkbox"].d2l-input-checkbox:indeterminate {
-			background-image: none;
-			position: relative;
-		}
-		input[type="checkbox"].d2l-input-checkbox:checked::after,
-		input[type="checkbox"].d2l-input-checkbox:indeterminate::after {
-			background-color: FieldText;
-			content: "";
-			display: block;
-			height: ${cssSizes.inputBoxSize}rem;
-			left: 50%;
-			position: absolute;
-			top: 50%;
-			transform: translate(-50%, -50%);
-			width: ${cssSizes.inputBoxSize}rem;
-		}
-
-		input[type="checkbox"].d2l-input-checkbox:disabled,
-		input[type="checkbox"].d2l-input-checkbox:where([aria-disabled="true"]) {
-			opacity: var(--d2l-theme-opacity-disabled-control);
-		}
-
-		input[type="checkbox"].d2l-input-checkbox:checked::after {
-			mask-image: var(--d2l-input-checkbox-check-image);
-		}
-
-		input[type="checkbox"].d2l-input-checkbox:indeterminate::after {
-			mask-image: var(--d2l-input-checkbox-indeterminate-image);
-		}
-	}
-`;
+export const checkboxStyles = _generateInputCheckboxStyles('input[type="checkbox"].d2l-input-checkbox');
 
 /**
  * A component that can be used to show a checkbox and optional visible label.
